Support remote scripts & stylesheets referenced within local HTML files
Previously, when Imagely was used to render local HTML files, those files could only reference local scripts or stylesheets. Any remote scripts or stylesheets (i.e. with a URL src or href) would not be properly loaded and executed.

Now, external scripts and stylesheets are fetched and inlined regardless of whether they're a local file or a remote URL.